feat: add Delete Confirmations

+++ client/src/main/java/org/admin/UI/window/enterpriseWindow/dialog/infos/AdminInfoDialog.java
@@ -5,10 +5,7 @@ import javafx.geometry.Pos;
 import javafx.geometry.VPos;
 import javafx.scene.Node;
 import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.control.TextArea;
-import javafx.scene.control.TextField;
+import javafx.scene.control.*;
 import javafx.scene.layout.*;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
@@ -18,6 +15,8 @@ import org.admin.connection.getRequests.GetUser;
 import org.admin.controller.AdminController;
 import org.admin.model.Response;
 import org.admin.model.User;
+
+import java.util.Optional;
 
 public class AdminInfoDialog extends Main {
     public static void show(Long id, Node node){
@@ -109,7 +108,7 @@ public class AdminInfoDialog extends Main {
         cancelButton.setOnAction(event -> dialog.close());
 
         deleteButton.setOnAction(event -> {
-            //TODO ALERT DELETE
+            if(!showDeleteAdminConfirmation()) return;
             Response response = DeleteUser.deleteById(token, admin.getId());
             if(response.getCode() == 200){
                 dialog.close();
@@ -123,5 +122,23 @@ public class AdminInfoDialog extends Main {
 
         dialog.setScene(dialogScene);
         dialog.showAndWait();
+    }
+
+    public static boolean showDeleteAdminConfirmation() {
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Предупреждение");
+        alert.setHeaderText("Удаление администратора");
+        alert.setContentText("Вы уверены, что хотите безвозвратно удалить администратора?");
+
+        // Настраиваем кнопки
+        ButtonType buttonTypeOk = new ButtonType("OK", ButtonBar.ButtonData.OK_DONE);
+        ButtonType buttonTypeCancel = new ButtonType("Отмена", ButtonBar.ButtonData.CANCEL_CLOSE);
+        alert.getButtonTypes().setAll(buttonTypeOk, buttonTypeCancel);
+
+        // Ждём выбора пользователя
+        Optional<ButtonType> result = alert.showAndWait();
+
+        // Возвращаем true, если нажата OK
+        return result.isPresent() && result.get() == buttonTypeOk;
     }
 }
